LayoutTests/fast/exclusions/parsing-wrap-shape-inside-expected.txt

Testing the parsing of the -webkit-shape-inside property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testCSSText("auto") is "auto"
PASS testComputedStyle("auto") is "auto"
PASS testCSSText("outside-shape") is "outside-shape"
PASS testComputedStyle("outside-shape") is "outside-shape"
PASS testCSSText("rectangle(10px, 20px, 30px, 40px)") is "rectangle(10px, 20px, 30px, 40px)"
PASS testComputedStyle("rectangle(10px, 20px, 30px, 40px)") is "rectangle(10px, 20px, 30px, 40px)"
PASS testCSSText("rectangle(10px, 20px, 30px, 40px, 5px)") is "rectangle(10px, 20px, 30px, 40px, 5px)"
PASS testComputedStyle("rectangle(10px, 20px, 30px, 40px, 5px)") is "rectangle(10px, 20px, 30px, 40px, 5px)"
PASS testCSSText("rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS testComputedStyle("rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS testCSSText("circle(10px, 20px, 30px)") is "circle(10px, 20px, 30px)"
PASS testComputedStyle("circle(10px, 20px, 30px)") is "circle(10px, 20px, 30px)"
PASS testCSSText("ellipse(10px, 20px, 30px, 40px)") is "ellipse(10px, 20px, 30px, 40px)"
PASS testComputedStyle("ellipse(10px, 20px, 30px, 40px)") is "ellipse(10px, 20px, 30px, 40px)"
PASS testCSSText("polygon(10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS testComputedStyle("polygon(10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS testCSSText("polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)") is "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)"
PASS testComputedStyle("polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)") is "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)"
PASS testCSSText("polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS testComputedStyle("polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS testNotInherited("auto", "rectangle(10px, 20px, 30px, 40px)") is "parent: auto, child: rectangle(10px, 20px, 30px, 40px)"
PASS testNotInherited("outside-shape", "rectangle(10px, 20px, 30px, 40px)") is "parent: outside-shape, child: rectangle(10px, 20px, 30px, 40px)"
PASS testNotInherited("rectangle(10px, 20px, 30px, 40px)", "initial") is "parent: rectangle(10px, 20px, 30px, 40px), child: outside-shape"
PASS testNotInherited("rectangle(10px, 20px, 30px, 40px)", "") is "parent: rectangle(10px, 20px, 30px, 40px), child: outside-shape"
PASS testNotInherited("rectangle(10px, 20px, 30px, 40px)", "inherit") is "parent: rectangle(10px, 20px, 30px, 40px), child: rectangle(10px, 20px, 30px, 40px)"
PASS testNotInherited("", "inherit") is "parent: outside-shape, child: outside-shape"
PASS testNotInherited("auto", "inherit") is "parent: auto, child: auto"
PASS testCSSText("calc()") is ""
PASS testComputedStyle("calc()") is "outside-shape"
PASS testCSSText("none") is ""
PASS testComputedStyle("none") is "outside-shape"
PASS testCSSText("rectangle()") is ""
PASS testComputedStyle("rectangle()") is "outside-shape"
PASS testCSSText("rectangle(10px)") is ""
PASS testComputedStyle("rectangle(10px)") is "outside-shape"
PASS testCSSText("rectangle(10px, 10px)") is ""
PASS testComputedStyle("rectangle(10px, 10px)") is "outside-shape"
PASS testCSSText("rectangle(10px, 20px, 30px)") is ""
PASS testComputedStyle("rectangle(10px, 20px, 30px)") is "outside-shape"
PASS testCSSText("rectangle(10px 20px 30px 40px)") is ""
PASS testComputedStyle("rectangle(10px 20px 30px 40px)") is "outside-shape"
PASS testCSSText("rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is ""
PASS testComputedStyle("rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is "outside-shape"
PASS testCSSText("circle()") is ""
PASS testComputedStyle("circle()") is "outside-shape"
PASS testCSSText("circle(10px)") is ""
PASS testComputedStyle("circle(10px)") is "outside-shape"
PASS testCSSText("circle(10px, 20px)") is ""
PASS testComputedStyle("circle(10px, 20px)") is "outside-shape"
PASS testCSSText("circle(10px 20px 30px)") is ""
PASS testComputedStyle("circle(10px 20px 30px)") is "outside-shape"
PASS testCSSText("circle(10px, 20px, 30px, 40px)") is ""
PASS testComputedStyle("circle(10px, 20px, 30px, 40px)") is "outside-shape"
PASS testCSSText("ellipse()") is ""
PASS testComputedStyle("ellipse()") is "outside-shape"
PASS testCSSText("ellipse(10px)") is ""
PASS testComputedStyle("ellipse(10px)") is "outside-shape"
PASS testCSSText("ellipse(10px, 20px)") is ""
PASS testComputedStyle("ellipse(10px, 20px)") is "outside-shape"
PASS testCSSText("ellipse(10px, 20px, 30px)") is ""
PASS testComputedStyle("ellipse(10px, 20px, 30px)") is "outside-shape"
PASS testCSSText("ellipse(10px 20px 30px 40px)") is ""
PASS testComputedStyle("ellipse(10px 20px 30px 40px)") is "outside-shape"
PASS testCSSText("polygon()") is ""
PASS testComputedStyle("polygon()") is "outside-shape"
PASS testCSSText("polygon(evenodd 10px 20px, 30px 40px, 40px 50px)") is ""
PASS testComputedStyle("polygon(evenodd 10px 20px, 30px 40px, 40px 50px)") is "outside-shape"
PASS testCSSText("polygon(nonzero 10px 20px, 30px 40px, 40px 50px)") is ""
PASS testComputedStyle("polygon(nonzero 10px 20px, 30px 40px, 40px 50px)") is "outside-shape"
PASS testCSSText("polygon(nonzero)") is ""
PASS testComputedStyle("polygon(nonzero)") is "outside-shape"
PASS testCSSText("polygon(evenodd)") is ""
PASS testComputedStyle("polygon(evenodd)") is "outside-shape"
PASS testCSSText("polygon(10px)") is ""
PASS testComputedStyle("polygon(10px)") is "outside-shape"
PASS testCSSText("polygon(nonzero,10px)") is ""
PASS testComputedStyle("polygon(nonzero,10px)") is "outside-shape"
PASS testCSSText("polygon(evenodd,12px)") is ""
PASS testComputedStyle("polygon(evenodd,12px)") is "outside-shape"
PASS testCSSText("polygon(10px, 20px, 30px, 40px, 40px, 50px)") is ""
PASS testComputedStyle("polygon(10px, 20px, 30px, 40px, 40px, 50px)") is "outside-shape"
PASS successfullyParsed is true

TEST COMPLETE
